separate intersection generation and intersection normalization
