+++ include/extractor/guidance/intersection_generator.hpp
@@ -6,9 +6,7 @@
 #include "extractor/guidance/intersection.hpp"
 #include "extractor/query_node.hpp"
 #include "extractor/restriction_map.hpp"
-#include "extractor/suffix_table.hpp"
 #include "util/attributes.hpp"
-#include "util/name_table.hpp"
 #include "util/node_based_graph.hpp"
 #include "util/typedefs.hpp"
 
@@ -25,7 +23,6 @@ namespace guidance
 // from it. For this all turn possibilities are analysed.
 // We consider turn restrictions to indicate possible turns. U-turns are generated based on profile
 // decisions.
-
 class IntersectionGenerator
 {
   public:
@@ -33,9 +30,7 @@ class IntersectionGenerator
                           const RestrictionMap &restriction_map,
                           const std::unordered_set<NodeID> &barrier_nodes,
                           const std::vector<QueryNode> &node_info_list,
-                          const CompressedEdgeContainer &compressed_edge_container,
-                          const util::NameTable &name_table,
-                          const SuffixTable &street_name_suffix_table);
+                          const CompressedEdgeContainer &compressed_edge_container);
 
     Intersection operator()(const NodeID nid, const EdgeID via_eid) const;
 
@@ -54,10 +49,8 @@ class IntersectionGenerator
     const CoordinateExtractor &GetCoordinateExtractor() const;
 
     // Check for restrictions/barriers and generate a list of valid and invalid turns present at
-    // the
-    // node reached
-    // from `from_node` via `via_eid`
-    // The resulting candidates have to be analysed for their actual instructions later on.
+    // the node reached from `from_node` via `via_eid`. The resulting candidates have to be analysed
+    // for their actual instructions later on.
     OSRM_ATTR_WARN_UNUSED
     Intersection GetConnectedRoads(const NodeID from_node, const EdgeID via_eid) const;
 
@@ -66,47 +59,9 @@ class IntersectionGenerator
     const RestrictionMap &restriction_map;
     const std::unordered_set<NodeID> &barrier_nodes;
     const std::vector<QueryNode> &node_info_list;
+
+    // own state, used to find the correct coordinates along a road
     const CoordinateExtractor coordinate_extractor;
-    const util::NameTable &name_table;
-    const SuffixTable &street_name_suffix_table;
-
-    // check if two indices in an intersection can be seen as a single road in the perceived
-    // intersection representation. See below for an example. Utility function for
-    // MergeSegregatedRoads
-    bool CanMerge(const NodeID intersection_node,
-                  const Intersection &intersection,
-                  std::size_t first_index,
-                  std::size_t second_index) const;
-
-    // Merge segregated roads to omit invalid turns in favor of treating segregated roads as
-    // one.
-    // This function combines roads the following way:
-    //
-    //     *                           *
-    //     *        is converted to    *
-    //   v   ^                         +
-    //   v   ^                         +
-    //
-    // The treatment results in a straight turn angle of 180º rather than a turn angle of approx
-    // 160
-    OSRM_ATTR_WARN_UNUSED
-    Intersection MergeSegregatedRoads(const NodeID intersection_node,
-                                      Intersection intersection) const;
-
-    // The counterpiece to mergeSegregatedRoads. While we can adjust roads that split up at the
-    // intersection itself, it can also happen that intersections are connected to joining roads.
-    //
-    //     *                           *
-    //     *        is converted to    *
-    //   v   a ---                     a ---
-    //   v   ^                         +
-    //   v   ^                         +
-    //       b
-    //
-    // for the local view of b at a.
-    OSRM_ATTR_WARN_UNUSED
-    Intersection AdjustForJoiningRoads(const NodeID node_at_intersection,
-                                       Intersection intersection) const;
 };
 
 } // namespace guidance
